requirements.txt
Django==1.9.2
django-filter==0.13.0
django-routers==0.2
djangorestframework==3.3.3
psycopg2==2.6.2
python-dateutil==2.5.2
pyttsx==1.1
pytz==2016.3
requests==2.9.1
urllib3==1.14
virtualenv==15.0.3
wget==3.2
